+++ controllers/user.js
@@ -265,12 +265,15 @@ const Update = async (req, res) => {
 
     if (req.file && existingUser.avatar) {
         let path = `./uploads/images/${existingUser.avatar}`;
-        try {
-            fs.unlinkSync(path)
-            //file removed
-        } catch (error) {
-            console.log(error);
-            return res.status(500).json({ success: false, message: error, error: error })
+        if (existingUser.avatar !== 'avatar.png' && fs.existsSync(path)) {
+
+            try {
+                fs.unlinkSync(path)
+                //file removed
+            } catch (error) {
+                console.log(error);
+                return res.status(500).json({ success: false, message: error, error: error })
+            }
         }
         existingUser.avatar = req.file.filename;
 
@@ -314,6 +317,28 @@ const DeleteUser = async (req, res) => {
 
     if (existingUser.avatar) {
         let path = `./uploads/images/${existingUser.avatar}`;
+        if (fs.existsSync(path)) {
+            try {
+                fs.unlinkSync(path)
+                //file removed
+            } catch (error) {
+                console.log(error);
+                return res.status(500).json({ success: false, message: error, error: error })
+            }
+        }
+    }
+    if (existingUser.cin) {
+        let path = `./uploads/images/${existingUser.cin}`;
+        try {
+            fs.unlinkSync(path)
+            //file removed
+        } catch (error) {
+            console.log(error);
+            return res.status(500).json({ success: false, message: error, error: error })
+        }
+    }
+    if (existingUser.patent) {
+        let path = `./uploads/images/${existingUser.patent}`;
         try {
             fs.unlinkSync(path)
             //file removed
@@ -548,12 +573,14 @@ const Update_Finance = async (req, res) => {
 
     if (req.file && existingUser.avatar) {
         let path = `./uploads/images/${existingUser.avatar}`;
-        try {
-            fs.unlinkSync(path)
-            //file removed
-        } catch (error) {
-            console.log(error);
-            return res.status(500).json({ success: false, message: error, error: error })
+        if (existingUser.avatar !== 'avatar.png' && fs.existsSync(path)) {
+            try {
+                fs.unlinkSync(path)
+                //file removed
+            } catch (error) {
+                console.log(error);
+                return res.status(500).json({ success: false, message: error, error: error })
+            }
         }
         existingUser.avatar = req.file.filename;
 
@@ -601,36 +628,43 @@ const Update_PDV = async (req, res) => {
 
     if (req.files && req.files.avatar && existingUser.avatar) {
         let path = `./uploads/images/${existingUser.avatar}`;
-        try {
-            fs.unlinkSync(path);
-            // File removed
-        } catch (error) {
-            console.log(error);
-            return res.status(500).json({ success: false, message: error, error: error });
+        if (existingUser.avatar !== 'avatar.png' && fs.existsSync(path)) {
+
+            try {
+                fs.unlinkSync(path);
+                // File removed
+            } catch (error) {
+                console.log(error);
+                return res.status(500).json({ success: false, message: error, error: error });
+            }
         }
         existingUser.avatar = req.files.avatar[0].filename;
     }
 
     if (req.files && req.files.cin && existingUser.cin) {
-        let path = `./uploads/files/${existingUser.cin}`;
-        try {
-            fs.unlinkSync(path);
-            // File removed
-        } catch (error) {
-            console.log(error);
-            return res.status(500).json({ success: false, message: error, error: error });
+        let path = `./uploads/images/${existingUser.cin}`;
+        if (existingUser.cin !== 'cin.png' && fs.existsSync(path)) {
+            try {
+                fs.unlinkSync(path);
+                // File removed
+            } catch (error) {
+                console.log(error);
+                return res.status(500).json({ success: false, message: error, error: error });
+            }
         }
         existingUser.cin = req.files.cin[0].filename;
     }
 
     if (req.files && req.files.patent && existingUser.patent) {
-        let path = `./uploads/files/${existingUser.patent}`;
-        try {
-            fs.unlinkSync(path);
-            // File removed
-        } catch (error) {
-            console.log(error);
-            return res.status(500).json({ success: false, message: error, error: error });
+        let path = `./uploads/images/${existingUser.patent}`;
+        if (existingUser.patent !== 'patente.png' && fs.existsSync(path)) {
+            try {
+                fs.unlinkSync(path);
+                // File removed
+            } catch (error) {
+                console.log(error);
+                return res.status(500).json({ success: false, message: error, error: error });
+            }
         }
         existingUser.patent = req.files.patent[0].filename;
     }
@@ -644,9 +678,6 @@ const Update_PDV = async (req, res) => {
     existingUser.register_comm = register_comm;
     existingUser.shop_name = shop_name;
     existingUser.secter = secter;
-    existingUser.patent = patent;
-    existingUser.cin = cin;
-
     try {
         await existingUser.save();
     } catch (error) {
@@ -661,6 +692,61 @@ const Lock = async (req, res) => {
 
     const { lock } = req.body;
     const { id } = req.params;
+
+    if (lock === false) {
+        let existingUser;
+        try {
+            existingUser = await user.findById(id);
+        } catch (error) {
+            return res.status(500).json({ success: false, message: 'internalm error server', data: error });
+        }
+
+        if (!existingUser) {
+            return res.status(200).json({ success: false, message: 'user donst exist!!', data: null });
+        }
+
+
+
+        try {
+            await existingUser.deleteOne();
+        } catch (error) {
+            return res.status(400).json({ success: false, message: 'internal server error', data: error });
+        }
+
+        if (existingUser.avatar) {
+            let path = `./uploads/images/${existingUser.avatar}`;
+            if (fs.existsSync(path)) {
+                try {
+                    fs.unlinkSync(path)
+                    //file removed
+                } catch (error) {
+                    console.log(error);
+                    return res.status(500).json({ success: false, message: error, error: error })
+                }
+            }
+        }
+        if (existingUser.cin) {
+            let path = `./uploads/images/${existingUser.cin}`;
+            try {
+                fs.unlinkSync(path)
+                //file removed
+            } catch (error) {
+                console.log(error);
+                return res.status(500).json({ success: false, message: error, error: error })
+            }
+        }
+        if (existingUser.patent) {
+            let path = `./uploads/images/${existingUser.patent}`;
+            try {
+                fs.unlinkSync(path)
+                //file removed
+            } catch (error) {
+                console.log(error);
+                return res.status(500).json({ success: false, message: error, error: error })
+            }
+        }
+        return res.status(200).json({ success: true, message: 'user updated successfully', data: existingUser });
+    }
 
     let existingUser;
     try {
